An alternate, hopefully more reliable tty reading routine
On my system the color lookup response is
"\a\e]4;xxx;rgb:rrrr/gggg/bbbb" (max 27 characters, with each RGB
value duplicated). Since we only grab 23, this isn't enough and
many of the color lookups fail. Additionally the module causes the
TTY to spew some output to the user.

Changing 23 to 26 causes the tty to hang. So instead we replace
it with a loop that reads until the full response has been read.